Updates and Fixes
Fixes:
TtPoint Parcel writing from long to int

Updates:
Updated Strings to ints for DAL enums
Updated permission requests
Updated Parcelable interfaces
Updated Main Menu
Updated Gradle to auto increment build versions

+++ src/main/java/com/usda/fmsc/twotrails/objects/points/TtPoint.java
@@ -390,7 +390,7 @@ public abstract class TtPoint extends TtObject implements Comparable<TtPoint>, C
     public void writeToParcel(Parcel dest, int flags) {
         super.writeToParcel(dest, flags);
 
-        dest.writeLong(_Index);
+        dest.writeInt(_Index);
         dest.writeInt(_PID);
         dest.writeSerializable(_Time);
